Add OpInfo test to check that floating point inputs in OpInfos have requires_grad set to True
This test detected a number of sampling methods that were not generating
the samples as expected, e.g. `index_put`, `cosine_embedding`, `stft`, but
perhaps most notably the generator for `BinOps`.

It also detected that `reminder` and `fmod` did not have implemented the
backward formula for the second input. I added this in the previous PR.

ghstack-source-id: 85485b43826c54d46120f7289142b45dbd8aad4c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69909