[SR] Make fused_gather_ranges_to_dense work on graph outputs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71498

We previously disabled `FuseListUnpack` if the fused outputs of the op would alias the graph outputs. The concern was that some ops were assuming that `p_node->Output(0).isTensor()` implies `p_node->Output(i).isTensor()` for all `i > 0`. This condition can be violated if there exists both managed and unmanaged tensors in the output list.

Instead of adding this special case and missing out on some fusions, we should implement fused ops correctly.

Differential Revision: [D33663969](https://our.internmc.facebook.com/intern/diff/D33663969/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D33663969/)!
ghstack-source-id: 148641456