[package] fix orderedimporter dummy package check
The current check that we have for dummy packages is too expansive; it
will skip anything without a `__file__`, including extension modules in
the standard library.

So first check if a module was created by torch.package before skipping
it, which should rule out anything accidentally getting skipped (as the
only time torch.package creates something without a `__file__` is the
dummy case).

ghstack-source-id: 144013e80db4abe1f8b9415fafbe2ef4bdcd3b36
Pull Request resolved: https://github.com/pytorch/pytorch/pull/72533